requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
plotly==5.19.0
pyyaml==6.0.1
pydantic==2.4.2
python-dotenv==1.0.1
openai==1.12.0
joblib==1.3.2

# Testing dependencies
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
flake8>=7.0.0
black>=24.1.0
mypy>=1.8.0

# FastAPI dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
